move format diff to compare_logs to pretty print
